Atlassian contributed: Bamboo 10.x compatibility - fixing wiring
